refactor: refactor swap calculations to use int64 types
- Change the parameter type of `CalculateProgress` from `float64` to `int64`
- Update calculation in `CalculateProgress` to handle `int64` amount
- Modify associated test cases to use `int64` instead of `float64` for amounts
- Rename `SwapDetails` to `SwapDetail` and change its structure to use `int64` for token amounts
- Update the return type of `GetTotalSwapAmount` from `float64` to `int64`
- Adjust logic in `GetTotalSwapAmount` to sum amounts as `int64`
- Change the `swap_details` field in `Transaction` to a repeated field of `SwapDetail` type
- Update various import paths and module dependencies within the `go.mod` file
- Ensure proper handling of transaction receipts and swap event logs in `TransactionRepoImpl`

+++ entity/domain/core/biz/agg_task.go
@@ -41,12 +41,12 @@ func NewTask(
 }
 
 // CalculateProgress calculates the progress of a task based on the given amount.
-func (t *Task) CalculateProgress(amount float64) int {
+func (t *Task) CalculateProgress(amount int64) int {
 	if t.Criteria == nil {
 		return 0
 	}
 
-	progress := int(amount / t.Criteria.MinTransactionAmount * 100)
+	progress := int((float64(amount) / t.Criteria.MinTransactionAmount) * 100)
 	if progress > 100 {
 		progress = 100
 	}
